Improve performance of lookup table by using a comparison tree
Folded characters tend to be clustered in ranges based on bytes 1-2 and
byte 3 in the BE expansion of the bytes in the `char`'s `u32`
representation.

This change exploits that fact by allowing large chunks of the lookup
table to be skipped based on the leading two bytes.

This change is inspired by a similar optimization in unicase.