Optimize Mapping iterator
Reduce the size of the mapping `Iter` to `size_of::<u8>()` by adding a
`Mapping::Empty` variant and changing the mapping arity as `char`s are
consumed. This significantly simplifies the implementation and makes it
easy to implement `DoubleEndedIterator`.

This PR also implements `Mapping::len` which, combined with the above
change, significantly simplifies `<Iter as Iterator>::size_hint` and
`<Iter as Iterator>::count`.

This PR adds additional tests for `<Iter as DoubleEndedIterator>::next_back`
and builds on the tests for the mapping iterator added in GH-13.