Set crate minimum supported Rust version to 1.52.0
Rust 1.52.0 stabilized several char ASCII methods as const which are
required by `lookup` if we wish to make `lookup` const.

https://blog.rust-lang.org/2021/05/06/Rust-1.52.0.html#stabilized-apis

This commit adds a job to the CI GitHub Actions workflow to test
foccacia with the declared MSRV of 1.52.0. Due to several panics in
incremental compilation in Rust 1.52.0, this job sets the
`CARGO_INCREMENTAL` env variable to 0 to disable incremental
compilation.