Ensure crate abides by the Unicode Data Files and Software License
Update the SPDX license expression in `Cargo.toml` to be `MIT AND
Unicode-DFS-2016` to reflect the inclusion of `CaseFolding.txt` and the
generated Rust sources that are derived from it.

Embed a copy of https://www.unicode.org/license.txt in the repository as
`LICENSE-UNICODE` and distribute this license in crates.io bundles.

Update the License section in `README.md` and add explainer text to the
rake task for fetching the Unicode Data Files and the script that
generates source from the Unicode Data Files.

When updating the `CaseFolding.txt` data file with `rake unicode:update`,
also fetch the license from https://www.unicode.org/license.txt.